+++ app/src/main/java/com/example/todolist/view/MainFragment.kt
@@ -8,8 +8,7 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.example.todolist.R
-import com.example.todolist.view.composecomponents.MainScreen
+import com.example.todolist.view.screens.MainScreen
 import com.example.todolist.viewmodel.NotesViewModel
 import dagger.hilt.android.AndroidEntryPoint
